+++ muutils/json_serialize/serializable_dataclass.py
@@ -11,262 +11,10 @@ import warnings
 from typing import Any, Callable, Optional, Type, TypeVar, Union
 
 from muutils.validate_type import validate_type
+from muutils.json_serialize.serializable_field import SerializableField, serializable_field
+from muutils.json_serialize.util import array_safe_eq, dc_eq
 
 # pylint: disable=bad-mcs-classmethod-argument, too-many-arguments, protected-access
-
-
-class SerializableField(dataclasses.Field):
-    """extension of `dataclasses.Field` with additional serialization properties"""
-
-    __slots__ = (
-        # from dataclasses.Field.__slots__
-        "name",
-        "type",
-        "default",
-        "default_factory",
-        "repr",
-        "hash",
-        "init",
-        "compare",
-        "metadata",
-        "kw_only",
-        "_field_type",  # Private: not to be used by user code.
-        # new ones
-        "serialize",
-        "serialization_fn",
-        "loading_fn",
-        "assert_type",
-    )
-
-    def __init__(
-        self,
-        default: Union[Any, dataclasses._MISSING_TYPE] = dataclasses.MISSING,
-        default_factory: Union[
-            Callable[[], Any], dataclasses._MISSING_TYPE
-        ] = dataclasses.MISSING,
-        init: bool = True,
-        repr: bool = True,
-        hash: Optional[bool] = None,
-        compare: bool = True,
-        # TODO: add field for custom comparator (such as serializing)
-        metadata: Optional[types.MappingProxyType] = None,
-        kw_only: Union[bool, dataclasses._MISSING_TYPE] = dataclasses.MISSING,
-        serialize: bool = True,
-        serialization_fn: Optional[Callable[[Any], Any]] = None,
-        loading_fn: Optional[Callable[[Any], Any]] = None,
-        assert_type: bool = True,
-        # TODO: add field for custom type assertion
-    ):
-        # TODO: should we do this check, or assume the user knows what they are doing?
-        if init and not serialize:
-            raise ValueError("Cannot have init=True and serialize=False")
-
-        # need to assemble kwargs in this hacky way so as not to upset type checking
-        super_kwargs: dict[str, Any] = dict(
-            default=default,
-            default_factory=default_factory,
-            init=init,
-            repr=repr,
-            hash=hash,
-            compare=compare,
-            kw_only=kw_only,
-        )
-
-        if metadata is not None:
-            super_kwargs["metadata"] = metadata
-        else:
-            super_kwargs["metadata"] = types.MappingProxyType({})
-
-        # special check, kw_only is not supported in python <3.9 and `dataclasses.MISSING` is truthy
-        if sys.version_info < (3, 10):
-            if super_kwargs["kw_only"] == True:  # noqa: E712
-                raise ValueError("kw_only is not supported in python >=3.9")
-            else:
-                del super_kwargs["kw_only"]
-
-        # actually init the super class
-        super().__init__(**super_kwargs)  # type: ignore[call-arg]
-
-        # now init the new fields
-        self.serialize: bool = serialize
-        self.serialization_fn: Optional[Callable[[Any], Any]] = serialization_fn
-        self.loading_fn: Optional[Callable[[Any], Any]] = loading_fn
-        self.assert_type: bool = assert_type
-
-    @classmethod
-    def from_Field(cls, field: dataclasses.Field) -> "SerializableField":
-        """copy all values from a `dataclasses.Field` to new `SerializableField`"""
-        return cls(
-            default=field.default,
-            default_factory=field.default_factory,
-            init=field.init,
-            repr=field.repr,
-            hash=field.hash,
-            compare=field.compare,
-            metadata=field.metadata,
-            kw_only=getattr(field, "kw_only", dataclasses.MISSING),  # for python <3.9
-            serialize=field.repr,
-            serialization_fn=None,
-            loading_fn=None,
-        )
-
-
-# Step 2: Create a serializable_field function
-# no type hint to avoid confusing mypy
-def serializable_field(*args, **kwargs):  # -> SerializableField:
-    """Create a new SerializableField
-
-    note that if not using ZANJ, and you have a class inside a container, you MUST provide
-    `serialization_fn` and `loading_fn` to serialize and load the container.
-    ZANJ will automatically do this for you.
-
-    ```
-    default: Any | dataclasses._MISSING_TYPE = dataclasses.MISSING,
-    default_factory: Callable[[], Any]
-    | dataclasses._MISSING_TYPE = dataclasses.MISSING,
-    init: bool = True,
-    repr: bool = True,
-    hash: Optional[bool] = None,
-    compare: bool = True,
-    metadata: types.MappingProxyType | None = None,
-    kw_only: bool | dataclasses._MISSING_TYPE = dataclasses.MISSING,
-    serialize: bool = True,
-    serialization_fn: Optional[Callable[[Any], Any]] = None,
-    loading_fn: Optional[Callable[[Any], Any]] = None,
-    assert_type: bool = True,
-    ```
-    """
-    return SerializableField(*args, **kwargs)
-
-
-# credit to https://stackoverflow.com/questions/51743827/how-to-compare-equality-of-dataclasses-holding-numpy-ndarray-boola-b-raises
-def array_safe_eq(a: Any, b: Any) -> bool:
-    """check if two objects are equal, account for if numpy arrays or torch tensors"""
-    if a is b:
-        return True
-
-    if (
-        str(type(a)) == "<class 'numpy.ndarray'>"
-        and str(type(b)) == "<class 'numpy.ndarray'>"
-    ) or (
-        str(type(a)) == "<class 'torch.Tensor'>"
-        and str(type(b)) == "<class 'torch.Tensor'>"
-    ):
-        return (a == b).all()
-
-    if (
-        str(type(a)) == "<class 'pandas.core.frame.DataFrame'>"
-        and str(type(b)) == "<class 'pandas.core.frame.DataFrame'>"
-    ):
-        return a.equals(b)
-
-    if isinstance(a, typing.Sequence) and isinstance(b, typing.Sequence):
-        return len(a) == len(b) and all(array_safe_eq(a1, b1) for a1, b1 in zip(a, b))
-
-    if isinstance(a, (dict, typing.Mapping)) and isinstance(b, (dict, typing.Mapping)):
-        return len(a) == len(b) and all(
-            array_safe_eq(k1, k2) and array_safe_eq(a[k1], b[k2])
-            for k1, k2 in zip(a.keys(), b.keys())
-        )
-
-    try:
-        return bool(a == b)
-    except (TypeError, ValueError) as e:
-        warnings.warn(f"Cannot compare {a} and {b} for equality\n{e}")
-        return NotImplemented  # type: ignore[return-value]
-
-
-def dc_eq(
-    dc1,
-    dc2,
-    except_when_class_mismatch: bool = False,
-    false_when_class_mismatch: bool = True,
-    except_when_field_mismatch: bool = False,
-) -> bool:
-    """checks if two dataclasses which (might) hold numpy arrays are equal
-
-        # Parameters:
-        - `dc1`: the first dataclass
-        - `dc2`: the second dataclass
-        - `except_when_class_mismatch: bool`
-            if `True`, will throw `TypeError` if the classes are different.
-            if not, will return false by default or attempt to compare the fields if `false_when_class_mismatch` is `False`
-            (default: `False`)
-        - `false_when_class_mismatch: bool`
-            only relevant if `except_when_class_mismatch` is `False`.
-            if `True`, will return `False` if the classes are different.
-            if `False`, will attempt to compare the fields.
-        - `except_when_field_mismatch: bool`
-            only relevant if `except_when_class_mismatch` is `False` and `false_when_class_mismatch` is `False`.
-            if `True`, will throw `TypeError` if the fields are different.
-            (default: `True`)
-
-        # Returns:
-        - `bool`: True if the dataclasses are equal, False otherwise
-
-        # Raises:
-        - `TypeError`: if the dataclasses are of different classes
-        - `AttributeError`: if the dataclasses have different fields
-
-    ```
-              [START]
-                 ▼
-           ┌───────────┐  ┌─────────┐
-           │dc1 is dc2?├─►│ classes │
-           └──┬────────┘No│ match?  │
-      ────    │           ├─────────┤
-     (True)◄──┘Yes        │No       │Yes
-      ────                ▼         ▼
-          ┌────────────────┐ ┌────────────┐
-          │ except when    │ │ fields keys│
-          │ class mismatch?│ │ match?     │
-          ├───────────┬────┘ ├───────┬────┘
-          │Yes        │No    │No     │Yes
-          ▼           ▼      ▼       ▼
-     ───────────  ┌──────────┐  ┌────────┐
-    { raise     } │ except   │  │ field  │
-    { TypeError } │ when     │  │ values │
-     ───────────  │ field    │  │ match? │
-                  │ mismatch?│  ├────┬───┘
-                  ├───────┬──┘  │    │Yes
-                  │Yes    │No   │No  ▼
-                  ▼       ▼     │   ────
-     ───────────────     ─────  │  (True)
-    { raise         }   (False)◄┘   ────
-    { AttributeError}    ─────
-     ───────────────
-    ```
-
-    """
-    if dc1 is dc2:
-        return True
-
-    if dc1.__class__ is not dc2.__class__:
-        if except_when_class_mismatch:
-            # if the classes don't match, raise an error
-            raise TypeError(
-                f"Cannot compare dataclasses of different classes: `{dc1.__class__}` and `{dc2.__class__}`"
-            )
-        else:
-            dc1_fields: set = set([fld.name for fld in dataclasses.fields(dc1)])
-            dc2_fields: set = set([fld.name for fld in dataclasses.fields(dc2)])
-            fields_match: bool = set(dc1_fields) == set(dc2_fields)
-
-            if not fields_match:
-                # if the fields match, keep going
-                if except_when_field_mismatch:
-                    raise AttributeError(
-                        f"dataclasses {dc1} and {dc2} have different fields: `{dc1_fields}` and `{dc2_fields}`"
-                    )
-                else:
-                    return False
-
-    return all(
-        array_safe_eq(getattr(dc1, fld.name), getattr(dc2, fld.name))
-        for fld in dataclasses.fields(dc1)
-        if fld.compare
-    )
-
 
 T = TypeVar("T")
 
@@ -278,7 +26,7 @@ class ZanjMissingWarning(UserWarning):
 _zanj_loading_needs_import: bool = True
 
 
-def zanj_register_loader_serializable_dataclass(cls: Type[T]):
+def zanj_register_loader_serializable_dataclass(cls: typing.Type[T]):
     """Register a serializable dataclass with the ZANJ backport
